+++ libsolidity/analysis/PostTypeChecker.cpp
@@ -19,13 +19,16 @@
 #include <libsolidity/analysis/PostTypeChecker.h>
 
 #include <libsolidity/ast/AST.h>
+#include <libsolidity/ast/CallGraph.h>
 #include <libsolidity/interface/Version.h>
 #include <liblangutil/ErrorReporter.h>
 #include <liblangutil/SemVerHandler.h>
 #include <libsolutil/Algorithms.h>
 
 #include <boost/range/adaptor/map.hpp>
+#include <boost/range/join.hpp>
 #include <memory>
+#include <set>
 
 using namespace std;
 using namespace solidity;
@@ -362,6 +365,77 @@ private:
 	/// Flag indicating whether we are currently inside a StructDefinition.
 	int m_insideStruct = 0;
 };
+
+struct CircularContractDependencies: public PostTypeChecker::Checker
+{
+	CircularContractDependencies(ErrorReporter& _errorReporter):
+		Checker(_errorReporter)
+	{}
+
+	bool visit(ContractDefinition const& _contract) override
+	{
+		auto joined = boost::join(
+			_contract.annotation().creationCallGraph->get()->createdContracts,
+			_contract.annotation().deployedCallGraph->get()->createdContracts
+		);
+
+		for (auto const& [contract, referencee]: joined)
+			if (contract != &_contract)
+				_contract.annotation().contractDependencies.emplace(contract, referencee);
+
+		return true;
+	}
+
+	void endVisit(ContractDefinition const& _contract) override
+	{
+		auto const& [contract, referencee] = dependenciesAreCyclic(_contract);
+
+		if (contract)
+			m_foundCycles.emplace(contract, referencee);
+	}
+
+	void finalize() override
+	{
+		for (auto const& [cycle, referencee]: m_foundCycles)
+		{
+			SecondarySourceLocation ssl{};
+			ssl.append(string{"Referenced contract"}, cycle->location());
+
+			m_errorReporter.typeError(
+				7813_error,
+				referencee->location(),
+				ssl,
+				"Circular reference found."
+			);
+		}
+	}
+
+private:
+	map<ContractDefinition const*, ASTNode const*, ASTNode::CompareByID> m_foundCycles;
+
+
+	pair<ContractDefinition const*, ASTNode const*> dependenciesAreCyclic(
+		ContractDefinition const& _currentContract
+	) const
+	{
+		auto visitor = [&](ContractDefinition const& _contract, util::CycleDetector<ContractDefinition>& _cycleDetector, size_t _depth)
+		{
+			if (_depth >= 256)
+				m_errorReporter.fatalTypeError(7864_error, _contract.location(), "Contract dependencies exhausting cyclic dependency validator");
+
+			for (auto& [contract, referencee]: _contract.annotation().contractDependencies)
+				if (_cycleDetector.run(*contract))
+					return;
+		};
+
+		ContractDefinition const* loop = util::CycleDetector<ContractDefinition>(visitor).run(_currentContract);
+
+		if (loop)
+			return {loop, _currentContract.annotation().contractDependencies[loop]};
+
+		return {nullptr, nullptr};
+	}
+};
 }
 
 PostTypeChecker::PostTypeChecker(langutil::ErrorReporter& _errorReporter): m_errorReporter(_errorReporter)
@@ -371,4 +445,5 @@ PostTypeChecker::PostTypeChecker(langutil::ErrorReporter& _errorReporter): m_err
 	m_checkers.push_back(make_shared<ModifierContextChecker>(_errorReporter));
 	m_checkers.push_back(make_shared<EventOutsideEmitChecker>(_errorReporter));
 	m_checkers.push_back(make_shared<NoVariablesInInterfaceChecker>(_errorReporter));
+	m_checkers.push_back(make_shared<CircularContractDependencies>(_errorReporter));
 }