+++ libsolidity/analysis/TypeChecker.cpp
@@ -329,6 +329,12 @@ void TypeChecker::endVisit(ModifierDefinition const& _modifier)
 
 bool TypeChecker::visit(FunctionDefinition const& _function)
 {
+	if (_function.isFree())
+	{
+		solAssert(m_currentFreeFunction == nullptr, "");
+		m_currentFreeFunction = &_function;
+	}
+
 	if (_function.markedVirtual())
 	{
 		if (_function.isFree())
@@ -494,6 +500,15 @@ bool TypeChecker::visit(FunctionDefinition const& _function)
 		typeCheckConstructor(_function);
 
 	return false;
+}
+
+void TypeChecker::endVisit(FunctionDefinition const& _function)
+{
+	if (_function.isFree())
+	{
+		solAssert(m_currentFreeFunction == &_function, "");
+		m_currentFreeFunction = nullptr;
+	}
 }
 
 bool TypeChecker::visit(VariableDeclaration const& _variable)
@@ -2627,22 +2642,14 @@ void TypeChecker::endVisit(NewExpression const& _newExpression)
 		if (contract->abstract())
 			m_errorReporter.typeError(4614_error, _newExpression.location(), "Cannot instantiate an abstract contract.");
 
-		if (m_currentContract)
+		if (m_currentContract || m_currentFreeFunction)
 		{
-			// TODO this is not properly detecting creation-cycles if they go through
-			// internal library functions or free functions. It will be caught at
-			// code generation time, but it would of course be better to catch it here.
-			m_currentContract->annotation().contractDependencies.insert(contract);
 			solAssert(
 				!contract->annotation().linearizedBaseContracts.empty(),
 				"Linearized base contracts not yet available."
 			);
-			if (contractDependenciesAreCyclic(*m_currentContract))
-				m_errorReporter.typeError(
-					4579_error,
-					_newExpression.location(),
-					"Circular reference for contract creation (cannot create instance of derived or same contract)."
-				);
+
+			currentContractDependencies()->contractDependencies.emplace(contract, &_newExpression);
 		}
 
 		_newExpression.annotation().type = FunctionType::newExpressionType(*contract);
@@ -2918,21 +2925,9 @@ bool TypeChecker::visit(MemberAccess const& _memberAccess)
 					_memberAccess.location(),
 					"\"runtimeCode\" is not available for contracts containing immutable variables."
 				);
-			if (m_currentContract)
-			{
-				// TODO in the same way as with ``new``,
-				// this is not properly detecting creation-cycles if they go through
-				// internal library functions or free functions. It will be caught at
-				// code generation time, but it would of course be better to catch it here.
 
-				m_currentContract->annotation().contractDependencies.insert(&accessedContractType.contractDefinition());
-				if (contractDependenciesAreCyclic(*m_currentContract))
-					m_errorReporter.typeError(
-						4224_error,
-						_memberAccess.location(),
-						"Circular reference for contract code access."
-					);
-			}
+			if (m_currentContract || m_currentFreeFunction)
+				currentContractDependencies()->contractDependencies.emplace(&accessedContractType.contractDefinition(), &_memberAccess);
 		}
 		else if (magicType->kind() == MagicType::Kind::MetaType && memberName == "name")
 			annotation.isPure = true;
@@ -2950,6 +2945,11 @@ bool TypeChecker::visit(MemberAccess const& _memberAccess)
 				"\"chainid\" is not supported by the VM version."
 			);
 	}
+	else if (m_currentContract || m_currentFreeFunction)
+		if (auto typeType = dynamic_cast<TypeType const*>(exprType))
+			if (auto contractType = dynamic_cast<ContractType const*>(typeType->actualType()))
+				if (&contractType->contractDefinition() != m_currentContract)
+					currentContractDependencies()->contractDependencies.emplace(&contractType->contractDefinition(), &_memberAccess);
 
 	if (
 		_memberAccess.expression().annotation().type->category() == Type::Category::Address &&
@@ -3392,22 +3392,6 @@ void TypeChecker::endVisit(UsingForDirective const& _usingFor)
 			_usingFor.location(),
 			"The \"using for\" directive is not allowed inside interfaces."
 		);
-}
-
-bool TypeChecker::contractDependenciesAreCyclic(
-	ContractDefinition const& _contract,
-	std::set<ContractDefinition const*> const& _seenContracts
-) const
-{
-	// Naive depth-first search that remembers nodes already seen.
-	if (_seenContracts.count(&_contract))
-		return true;
-	set<ContractDefinition const*> seen(_seenContracts);
-	seen.insert(&_contract);
-	for (auto const* c: _contract.annotation().contractDependencies)
-		if (contractDependenciesAreCyclic(*c, seen))
-			return true;
-	return false;
 }
 
 Declaration const& TypeChecker::dereference(Identifier const& _identifier) const